feat: render blogs as cards refactor: rename Blog to Blogs so it follows the project naming convention
